Core/Instances: Instance lock rewrite (WIP)
Port From (https://github.com/TrinityCore/TrinityCore/commit/17665c929c3a9fb7fe75dd680648129bc1c1f874)

+++ Source/Game/Networking/Packets/CalendarPackets.cs
@@ -852,13 +852,13 @@ namespace Game.Networking.Packets
             data.WriteUInt64(InstanceID);
             data.WriteInt32(MapID);
             data.WriteUInt32(DifficultyID);
-            data.WriteUInt32((uint)ExpireTime);
+            data.WriteInt32(ExpireTime);
         }
 
         public ulong InstanceID;
         public int MapID;
         public uint DifficultyID;
-        public long ExpireTime;
+        public int ExpireTime;
     }
 
     struct CalendarSendCalendarEventInfo
